Check if installed or not

+++ install.sh
@@ -1,26 +1,47 @@
 #!/bin/sh
 
 # install emacs
-sudo apt-get install emacs23
+if ! [ -f /usr/bin/emacs23 ]; then
+    sudo apt-get install emacs23
+fi
 
 # install apel
-wget http://git.chise.org/elisp/dist/apel/apel-10.8.tar.gz
-tar zxvf apel-10.8.tar.gz
-cd apel-10.8
-make
-sudo make install
-cd ..
+if ! [ -f /usr/local/share/emacs/23.3/site-lisp/emu/poe.el ]; then
+    wget http://git.chise.org/elisp/dist/apel/apel-10.8.tar.gz
+    tar zxvf apel-10.8.tar.gz
+    cd apel-10.8
+    make
+    sudo make install
+    cd ..
+fi
 
 # install trr
-wget https://trr22.googlecode.com/files/trr22_0.99-5.tar.gz
-tar zxvf trr22_0.99-5.tar.gz
-cd trr22-0.99
-cp ../apel-10.8/*.el .
-make all
-sudo make install
+if ! [ -d /usr/share/emacs/site-lisp/trr22 ]; then
+    wget https://trr22.googlecode.com/files/trr22_0.99-5.tar.gz
+    tar zxvf trr22_0.99-5.tar.gz
+    cd trr22-0.99
+    cp ../apel-10.8/*.el .
+    make all
+    sudo make install
+else
+    echo """
+    TRR is already installed.
+    To use trr on emacs, please add lines below to your emacs config file. (ex. ~/.emacs.d/init.el or ~/.emacs)
+
+    (add-to-list 'load-path "/usr/share/emacs/site-lisp/trr22")
+    (add-to-list 'load-path "/usr/local/share/emacs/23.3/site-lisp/emu")
+    (autoload 'trr "/usr/share/emacs/site-lisp/trr22/trr" nil t)
+
+    Now you can play trr on your emacs by "M-x trr".
+    """
+    return
+fi
+
 
 # install & use nkf
-sudo apt-get install nkf
+if ! [ -f /usr/bin/nkf ]; then
+    sudo apt-get install nkf
+fi
 sudo nkf -w --overwrite /var/lib/trr22/CONTENTS
 
 # add score file
@@ -28,13 +49,14 @@ sudo touch /var/lib/trr22/record/SCORE-IC
 
 # add trr and apel lisp files to lisp-path
 cat <<-EOF
---
-Okay, the installation was successfully ended.
-Finally, please add lines below to your emacs config file. (ex. ~/.emacs.d/init.el or ~/.emacs)
 
-(add-to-list 'load-path "/usr/share/emacs/site-lisp/trr22")
-(add-to-list 'load-path "/usr/local/share/emacs/23.3/site-lisp/emu")
-(autoload 'trr "/usr/share/emacs/site-lisp/trr22/trr" nil t)
+    Okay, the installation was successfully ended.
+    Finally, please add lines below to your emacs config file. (ex. ~/.emacs.d/init.el or ~/.emacs)
 
-Now you can play trr on your emacs by "M-x trr".
+    (add-to-list 'load-path "/usr/share/emacs/site-lisp/trr22")
+    (add-to-list 'load-path "/usr/local/share/emacs/23.3/site-lisp/emu")
+    (autoload 'trr "/usr/share/emacs/site-lisp/trr22/trr" nil t)
+
+    Now you can play trr on your emacs by "M-x trr".
+
 EOF
